quota: move quota package out of graphdriver
