+++ daemon/graphdriver/vfs/quota_linux.go
@@ -1,7 +1,7 @@
 package vfs // import "github.com/docker/docker/daemon/graphdriver/vfs"
 
 import (
-	"github.com/docker/docker/daemon/graphdriver/quota"
+	"github.com/docker/docker/quota"
 	"github.com/sirupsen/logrus"
 )
