+++ daemon/graphdriver/vfs/quota_unsupported.go
@@ -2,7 +2,7 @@
 
 package vfs // import "github.com/docker/docker/daemon/graphdriver/vfs"
 
-import "github.com/docker/docker/daemon/graphdriver/quota"
+import "github.com/docker/docker/quota"
 
 type driverQuota struct {
 }
