+++ quota/types.go
@@ -1,4 +1,4 @@
-package quota // import "github.com/docker/docker/daemon/graphdriver/quota"
+package quota // import "github.com/docker/docker/quota"
 
 import "sync"
